Style changes, mostly weakening types
git-svn-id: https://svn.apache.org/repos/asf/mahout/trunk@1032979 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/mahout/clustering/minhash/HashFactory.java
@@ -65,7 +65,7 @@ public class HashFactory {
     @Override
     public int hash(byte[] bytes) {
       long hashValue = 31;
-      for (byte byteVal : bytes) {
+      for (long byteVal : bytes) {
         hashValue *= seedA * byteVal;
         hashValue += seedB;
       }
@@ -87,7 +87,7 @@ public class HashFactory {
     @Override
     public int hash(byte[] bytes) {
       long hashValue = 31;
-      for (byte byteVal : bytes) {
+      for (long byteVal : bytes) {
         hashValue *= seedA * (byteVal >> 4);
         hashValue += seedB * byteVal + seedC;
       }
